third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/dom-breakpoints-expected.txt

Tests DOM breakpoints.

Running: testInsertChild
Test that 'Subtree Modified' breakpoint is hit when appending a child.
Set 'Subtree Modified' DOM breakpoint on rootElement.
Append childElement to rootElement.
Script execution paused.
Call stack:
    0) appendElement (dom-breakpoints.html:5)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on subtree modificationsdiv#rootElementChild div#rootElement added
Script execution resumed.

Running: testBreakpointToggle
Test that DOM breakpoint toggles properly using checkbox.
Set DOM breakpoint.
Uncheck DOM breakpoint.
DOM breakpoint should not be hit when disabled.
Check DOM breakpoint.
Test that DOM breakpoint is hit when re-enabled.
Script execution paused.
Call stack:
    0) modifyAttribute (dom-breakpoints.html:10)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on attribute modificationsdiv#rootElement
Script execution resumed.

Running: testInsertGrandchild
Test that 'Subtree Modified' breakpoint is hit when appending a grandchild.
Append grandchildElement to childElement.
Script execution paused.
Call stack:
    0) appendElement (dom-breakpoints.html:5)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on subtree modificationsdiv#rootElementDescendant div#childElement added
Script execution resumed.

Running: testRemoveChild
Test that 'Subtree Modified' breakpoint is hit when removing a child.
Remove grandchildElement.
Script execution paused.
Call stack:
    0) removeElement (dom-breakpoints.html:32)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on subtree modificationsdiv#rootElementDescendant div#grandchildElement removed
Script execution resumed.

Running: testInnerHTML
Test that 'Subtree Modified' breakpoint is hit exactly once when setting innerHTML.
Set childElement.innerHTML.
Script execution paused.
Call stack:
    0) setInnerHTML (dom-breakpoints.html:37)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on subtree modificationsdiv#rootElementDescendant div#childElement added
Script execution resumed.
Call breakDebugger, expect it to show up in next stack trace.
Script execution paused.
Call stack:
    0) breakDebugger (dom-breakpoints.html:41)
    1)  (:1)
Script execution resumed.

Running: testModifyAttribute
Test that 'Attribute Modified' breakpoint is hit when modifying attribute.
Set 'Attribute Modified' DOM breakpoint on rootElement.
Modify rootElement data-test attribute.
Script execution paused.
Call stack:
    0) modifyAttribute (dom-breakpoints.html:10)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on attribute modificationsdiv#rootElement
Script execution resumed.

Running: testModifyAttrNode
Test that 'Attribute Modified' breakpoint is hit when modifying Attr node.
Set 'Attribute Modified' DOM breakpoint on rootElement.
Modify rootElement data-test attribute.
Script execution paused.
Call stack:
    0) modifyAttrNode (dom-breakpoints.html:15)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on attribute modificationsdiv#rootElement
Script execution resumed.

Running: testSetAttrNode
Test that 'Attribute Modified' breakpoint is hit when adding a new Attr node.
Set 'Attribute Modified' DOM breakpoint on rootElement.
Modify rootElement data-foo attribute.
Script execution paused.
Call stack:
    0) setAttrNode (dom-breakpoints.html:22)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on attribute modificationsdiv#rootElement
Script execution resumed.

Running: testModifyStyleAttribute
Test that 'Attribute Modified' breakpoint is hit when modifying style attribute.
Set 'Attribute Modified' DOM breakpoint on rootElement.
Modify rootElement style.color attribute.
Script execution paused.
Call stack:
    0) modifyStyleAttribute (dom-breakpoints.html:27)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on attribute modificationsdiv#rootElement
Script execution resumed.

Running: testRemoveNode
Test that 'Node Removed' breakpoint is hit when removing a node.
Set 'Node Removed' DOM breakpoint on elementToRemove.
Remove elementToRemove.
Script execution paused.
Call stack:
    0) removeElement (dom-breakpoints.html:32)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on node removaldiv#elementToRemove
Script execution resumed.

Running: testReload
Test that DOM breakpoints are persisted between page reloads.
Set 'Subtree Modified' DOM breakpoint on rootElement.
Page reloaded.
Append childElement to rootElement.
Script execution paused.
Call stack:
    0) appendElement (dom-breakpoints.html:5)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on subtree modificationsdiv#rootElementChild div#rootElement added
Script execution resumed.

Running: testInsertChildIntoAuthorShadowTree
Test that 'Subtree Modified' breakpoint on author shadow root is hit when appending a child.
Set 'Subtree Modified' DOM breakpoint on author shadow root.
Append childElement to author shadow root.
Script execution paused.
Call stack:
    0) appendElementToOpenShadowRoot (dom-breakpoints.html:61)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on subtree modifications#shadow-root (open)Child #shadow-root (open) added
Script execution resumed.

Running: testReloadWithShadowElementBreakpoint
Test that shadow DOM breakpoints are persisted between page reloads.
Set 'Subtree Modified' DOM breakpoint on outerElement.
Page reloaded.
Append childElement to outerElement.
Script execution paused.
Call stack:
    0) appendElementToAuthorShadowTree (dom-breakpoints.html:55)
    1)  (:1)
    [setTimeout]
    0)  (:1)
Paused on subtree modificationsdiv#outerElementChild div#outerElement added
Script execution resumed.
